Implement CommunityTecAdapter with specific data transformation pipelines and remove obsolete CommunityTpflAdapter
